moved input away from main
