+++ main.js
@@ -5,10 +5,10 @@ import { Circle } from './game/shapes/circle.js';
 import { RigidBody } from './game/rigidBody.js';
 import { BALL_COLORS, BALL_RADIUS, BALL_SPEED, MAX_BALLS, TIME_STEP } from './settings.js';
 import { GameWorld } from './game/world.js';
-
+import { inputState, initInputListeners } from './game/input.js';
 
 // Initialize the canvas and set up drawing functionalities
-const canvas = document.getElementById('drawingCanvas');
+export const canvas = document.getElementById('drawingCanvas');
 const ctx = canvas.getContext('2d');
 
 // Set canvas dimensions
@@ -16,7 +16,7 @@ canvas.width = window.innerWidth;
 canvas.height = window.innerHeight;
 
 // world instance
-const world = new GameWorld();
+export const world = new GameWorld();
 
 // Detect which page we're on
 const isTesting = window.location.pathname.endsWith('test-engine.html');
@@ -32,57 +32,7 @@ if (!isTesting) {
     }
 }
 
-let mousePosition = new Vector2();
-
-// Mouse object to store position, velocity, and button states
-let mouse = {
-    position: new Vector2(),
-    lastPosition: new Vector2(),
-    velocity: new Vector2(),
-    leftDown: false,
-    rightDown: false
-};
-
-canvas.addEventListener('mousemove', (e) => {
-    const rect = canvas.getBoundingClientRect();
-    mouse.lastPosition.x = mouse.position.x;
-    mouse.lastPosition.y = mouse.position.y;
-    mouse.position.x = e.clientX - rect.left;
-    mouse.position.y = e.clientY - rect.top;
-    mouse.velocity = mouse.position.subtract(mouse.lastPosition);
-});
-
-canvas.addEventListener('mousedown', (e) => {
-    if (e.button === 0) mouse.leftDown = true;
-    if (e.button === 2) mouse.rightDown = true;
-
-    if (e.button === 0) { // Left mouse button
-        const color = BALL_COLORS[Math.floor(Math.random() * BALL_COLORS.length)];
-        world.createRigidBody(new Circle(new Vector2(mouse.position.x, mouse.position.y), BALL_RADIUS), null, color, new Vector2(0, 0));
-    }
-});
-
-canvas.addEventListener('mouseup', (e) => {
-    if (e.button === 0) mouse.leftDown = false;
-    if (e.button === 2) mouse.rightDown = false;
-});
-
-// Optional: prevent context menu on right click
-canvas.addEventListener('contextmenu', (e) => e.preventDefault());
-
-function resizeCanvas() {
-    canvas.width = window.innerWidth;
-    canvas.height = window.innerHeight;
-    // Do NOT reposition existing balls on resize
-}
-
-window.addEventListener('resize', () => {
-    resizeCanvas();
-});
-
-// Initial canvas setup
-resizeCanvas();
-
+initInputListeners();
 
 function drawWelcomeText() {
     ctx.fillStyle = 'black';
@@ -118,8 +68,8 @@ function animateTestEngine() {
     ctx.fillText('Physics Engine Test Animation', canvas.width / 2, 50);
 
     // Check closest ball info while right mouse button is held
-    if (mouse.rightDown) {
-        const { ball, distance, inside } = world.getClosestBallInfo(mouse.position);
+    if (inputState.mouse.rightDown) {
+        const { ball, distance, inside } = world.getClosestBallInfo(inputState.mouse.position);
         if (ball) {
             ctx.font = '20px Arial';
             ctx.fillStyle = inside ? 'green' : 'red';
@@ -129,10 +79,10 @@ function animateTestEngine() {
                 canvas.width / 2,
                 90
             );
-            // Move the ball to the mouse position if inside
+            // Move the ball to the inputState.mouse position if inside
             if (inside) {
-                ball.shape.position.x = mouse.position.x;
-                ball.shape.position.y = mouse.position.y;
+                ball.shape.position.x = inputState.mouse.position.x;
+                ball.shape.position.y = inputState.mouse.position.y;
             }
         } else {
             ctx.font = '20px Arial';
@@ -150,8 +100,6 @@ if (!isTesting) {
 } else {
     animateTestEngine();
 }
-
-
 
 
 // Example: spawn a new ball every 2 seconds (optional)
